hp-fe-69 - add composable formErrors  for login and registration
